Wait for tx to be in the mempool in test helper
Some tests that use the mineSuccessfulTransaction helper
sometimes fail in the CI because the `hardhat_mine` method
is sent before the transaction is actually in the mempool.

This seems to be caused by ethers returning a transaction
object before the request actually resolves (otherwise we
know that the tx would be in the mempool because hardhat
doesn't respond until it fully processed the transaction).

To fix that, we first send a no-op RPC method, and then
we also assert that the transaction is actually ready
to be mined.